Handle pre ios 14 images
